upgrade to lastest from GitHub. Fixes problem where threadlocal variables were not being freed on tread exit (requires new destructor argument on BeThreadLocalStorage::Create)

+++ iModelCore/libsrc/facebook/folly/MacAddress.cpp
@@ -16,6 +16,8 @@
 
 #include <folly/MacAddress.h>
 
+#include <ostream>
+
 #include <folly/Exception.h>
 #include <folly/IPAddressV6.h>
 
@@ -126,7 +128,7 @@ void MacAddress::parse(StringPiece str) {
     }
 
     // Update parsed with the newly parsed byte
-    parsed[byteIndex] = ((upper << 4) | lower);
+    parsed[byteIndex] = uint8_t((upper << 4) | lower);
   }
 
   if (p != str.end()) {
